Add default value for remote submission
Previously, the Y in [Yn] was capitalised, which normally indicates that Y is the default value - but it wasn't. This makes 'yes' the default value when the user is asked if they wish to submit their bropage, allowing them to press enter, rather than having to provide an answer to the question.